fix create bug

+++ server/src/ticTacToe/ticTacToeBot.gateway.ts
@@ -39,24 +39,52 @@ export class TicTacToeBotGateway {
 
       private async isPlaying(userId: string) {
             const isPlaying = await this.ticTacToeCommonService.isPlaying(userId);
-            if (isPlaying)
-                  return this.socketServer().socketEmitToRoomError('BadRequestException', userId, {
-                        details: { message: { type: 'game.already-join-other' } },
-                  });
+            if (isPlaying) {
+                  this.socketServer().socketEmitToRoomError(
+                        'BadRequestException',
+                        userId,
+                        {
+                              details: { message: { type: 'game.already-join-other' } },
+                        },
+                        'user',
+                  );
+                  return false;
+            }
+            return true;
       }
 
-      private async getGameFromCache(roomId: string) {
+      private async getGameFromCache(roomId: string, userId: string) {
             const game = await this.ticTacToeCommonService.getBoard(roomId);
-            if (!game) throw ioResponse.sendError({ details: { roomId: { type: 'user.not-found' } } }, 'NotFoundException');
+            if (!game) {
+                  this.socketServer().socketEmitToRoomError('NotFoundException', userId, { details: { roomId: { type: 'user.not-found' } } }, 'user');
+                  return null;
+            }
+
             return game;
       }
 
-      private isOwner(game: TicTacToeBoard, client: SocketExtend) {
-            const isOwner = game.users.find((item) => item.id === client.user.id);
-            if (!isOwner)
-                  return this.socketServer().socketEmitToRoomError('UnauthorizedException', client.user.id, {
-                        details: { roomId: { type: 'user.auth-failed' } },
-                  });
+      private isOwner(game: TicTacToeBoard, userId: string) {
+            const isOwner = game.users.find((item) => item.id === userId);
+            if (!isOwner) {
+                  this.socketServer().socketEmitToRoomError(
+                        'UnauthorizedException',
+                        userId,
+                        {
+                              details: { roomId: { type: 'user.auth-failed' } },
+                        },
+                        'user',
+                  );
+                  return false;
+            }
+            return true;
+      }
+
+      @UseGuards(UserSocketGuard)
+      @SubscribeMessage(TTTBotAction.TTT_BOT_CONNECT)
+      async handleConnectUser(@ConnectedSocket() client: SocketExtend) {
+            await client.join(`user-${client.user.id}`);
+
+            return this.socketServer().socketEmitToRoom(TTTBotAction.TTT_BOT_CONNECT, client.user.id, { data: {} }, 'user');
       }
 
       @UseGuards(UserSocketGuard)
@@ -96,14 +124,22 @@ export class TicTacToeBotGateway {
       ) {
             await this.isPlaying(client.user.id);
 
-            const getCacheGame = await this.getGameFromCache(body.roomId);
-            this.isOwner(getCacheGame, client);
+            const getCacheGame = await this.getGameFromCache(body.roomId, client.user.id);
+            if (!getCacheGame) return;
+
+            const isOwner = this.isOwner(getCacheGame, client.user.id);
+            if (!isOwner) return;
 
             const isStart = await this.ticTacToeService.startGame(body.roomId, client.user);
             if (!isStart)
-                  return this.socketServer().socketEmitToRoomError('BadRequestException', client.user.id, {
-                        details: { message: { type: 'game.wait-ready-player' } },
-                  });
+                  return this.socketServer().socketEmitToRoomError(
+                        'BadRequestException',
+                        client.user.id,
+                        {
+                              details: { message: { type: 'game.wait-ready-player' } },
+                        },
+                        'user',
+                  );
 
             return this.socketServer().socketEmitToRoom(TTTBotAction.TTT_BOT_START, getCacheGame.info.id, {}, 'tic-tac-toe-bot');
       }
@@ -115,21 +151,36 @@ export class TicTacToeBotGateway {
             client: SocketExtend,
             @MessageBody(new SocketJoiValidatorPipe(vAddMoveDto)) body: AddMoveDto,
       ) {
-            const getCacheGame = await this.getGameFromCache(body.roomId);
-            this.isOwner(getCacheGame, client);
+            const getCacheGame = await this.getGameFromCache(body.roomId, client.user.id);
+            if (!getCacheGame) return;
+
+            const isOwner = this.isOwner(getCacheGame, client.user.id);
+            if (!isOwner) return;
 
             if (getCacheGame.board[body.x][body.y] !== -1)
-                  return this.socketServer().socketEmitToRoomError('BadRequestException', client.user.id, {
-                        details: { message: { type: 'game.already-chose' } },
-                  });
+                  return this.socketServer().socketEmitToRoomError(
+                        'BadRequestException',
+                        client.user.id,
+                        {
+                              details: { message: { type: 'game.already-chose' } },
+                        },
+                        'user',
+                  );
 
             const isAddMove = await this.ticTacToeService.addMoveToBoard(body.roomId, client.user, body.x, body.y);
             if (!isAddMove)
-                  return this.socketServer().socketEmitToRoomError('BadRequestException', client.user.id, {
-                        details: { message: { type: 'game.wrong-turn' } },
-                  });
+                  return this.socketServer().socketEmitToRoomError(
+                        'BadRequestException',
+                        client.user.id,
+                        {
+                              details: { message: { type: 'game.wrong-turn' } },
+                        },
+                        'user',
+                  );
 
-            const getUpdateCacheGame = await this.getGameFromCache(body.roomId);
+            const getUpdateCacheGame = await this.getGameFromCache(body.roomId, client.user.id);
+            if (!getUpdateCacheGame) return;
+
             const userMove = await this.ticTacToeBotService.findBestMove(getUpdateCacheGame.board, 0);
             const botMove = await this.ticTacToeBotService.findBestMove(getUpdateCacheGame.board, 1);
 
@@ -150,8 +201,11 @@ export class TicTacToeBotGateway {
             @MessageBody(new SocketJoiValidatorPipe(vRoomIdDto))
             body: RoomIdDTO,
       ) {
-            const getCacheGame = await this.getGameFromCache(body.roomId);
-            this.isOwner(getCacheGame, client);
+            const getCacheGame = await this.getGameFromCache(body.roomId, client.user.id);
+            if (!getCacheGame) return;
+
+            const isOwner = this.isOwner(getCacheGame, client.user.id);
+            if (!isOwner) return;
 
             await this.ticTacToeCommonService.deleteBoard(body.roomId);
 
@@ -161,8 +215,11 @@ export class TicTacToeBotGateway {
       @UseGuards(UserSocketGuard)
       @SubscribeMessage(TTTBotAction.TTT_BOT_GET)
       async handleGetGame(@ConnectedSocket() client: SocketExtend, @MessageBody(new SocketJoiValidatorPipe(vRoomIdDto)) body: RoomIdDTO) {
-            const getCacheGame = await this.getGameFromCache(body.roomId);
-            this.isOwner(getCacheGame, client);
+            const getCacheGame = await this.getGameFromCache(body.roomId, client.user.id);
+            if (!getCacheGame) return;
+
+            const isOwner = this.isOwner(getCacheGame, client.user.id);
+            if (!isOwner) return;
 
             return this.socketServer().socketEmitToRoom(TTTBotAction.TTT_BOT_GET, getCacheGame.info.id, { data: getCacheGame }, 'tic-tac-toe-bot');
       }
